Better directory split of processing module
