+++ src/data/stats.test.ts
@@ -1,4 +1,4 @@
-import { Feature } from 'src/processing/types';
+import { Feature } from '../types';
 import { getTestSuiteStats } from './stats';
 
 const ThreePassedScenarios: Feature[] = [
